requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
bs4
arxiv
pymupdf
langchain-text-splitters
langchain_openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_community
langchain_chroma
langchain-openai
langchain-groq
langchain_groq
huggingface_hub
transformers 